+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 * Make the meta title format multi language aware ([#13](https://github.com/wanze/SeoMaestro/issues/13))
+* Update recommended meta description length  ([#12](https://github.com/wanze/SeoMaestro/issues/12))
 
 ## [0.8.0] - 2019-06-15